requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.27
python-dateutil==2.8.2
mysqlclient==2.2.4
